refactor: 상품 목록 API 호출 로직을 apis/product.ts로 분리
